Add START_ROW functionality to LongDivision

+++ belugapad/LongDivision.h
@@ -67,7 +67,7 @@
     int previousNumberPos;
     float currentTotal;
     float rowMultiplier;
-    float startRow;
+    int startRow;
     
     int currentTouchCount;
